extract curand error checking into curand_utils.h
